# pos-web/Swagger/pos-api-request.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "POS.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/PurchaseOrders": {
      "get": {
        "tags": [
          "PurchaseOrders"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PurchaseOrders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.CreatePurchaseOrderDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.CreatePurchaseOrderDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.CreatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
    "/api/PurchaseOrders/{id}": {
      "get": {
        "tags": [
          "PurchaseOrders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PurchaseOrders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.UpdatePurchaseOrderDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.UpdatePurchaseOrderDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.UpdatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.PurchaseOrder.PurchaseOrderDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/SalePayments": {
      "post": {
        "tags": [
          "SalePayments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.CreateSalePaymentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.CreateSalePaymentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.CreateSalePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.SalePaymentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.SalePaymentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.SalePaymentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/SalePayments/sale/{saleId}": {
      "get": {
        "tags": [
          "SalePayments"
        ],
        "parameters": [
          {
            "name": "saleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.SalePaymentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.SalePaymentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/POS.Application.DTOs.SalePayment.SalePaymentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Sales/{id}": {
      "get": {
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Sale.SaleDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Sale.SaleDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Sale.SaleDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "POS.Application.DTOs.PurchaseOrder.CreatePurchaseOrderDetailDto": {
        "type": "object",
        "properties": {
          "productVariantId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitCost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "POS.Application.DTOs.PurchaseOrder.CreatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedDeliveryDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.CreatePurchaseOrderDetailDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "POS.Application.DTOs.PurchaseOrder.UpdatePurchaseOrderDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "productVariantId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitCost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "POS.Application.DTOs.PurchaseOrder.UpdatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "expectedDeliveryDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/POS.Application.DTOs.PurchaseOrder.UpdatePurchaseOrderDetailDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "POS.Application.DTOs.SalePayment.CreateSalePaymentDto": {
        "type": "object",
        "properties": {
          "saleId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentMethodId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "POS.Application.DTOs.SalePayment.SalePaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "saleId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentMethodId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentMethodName": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your valid token in the text input below.\r\n\r\nExample: \"Bearer eyJhbGciOiJIUzI1Ni...\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}